// src/app/models/movie.ts
export interface Movie {
  imdbID: string;
  title: string;
  year: string;
  poster: string;
  genre?: string;
  director?: string;
  actors?: string;
  plot?: string;
  type?: string;
  imdbRating?: string;
  runtime?: string;
  rated?: string;
  released?: string;
  country?: string;
  language?: string;
  website?: string;
  boxOffice?: string;
  production?: string;
  awards?: string;
  metascore?: string;
  ratings?: string;
  response?: string;
  writer?: string;
  imdbVotes?: string;
  totalSeasons?: string;
}